Move progress_bar and json read/write to utils
